try to reject tests when GPU does not support
