+++ base/test/ccnppi_tests.cpp
@@ -13,9 +13,11 @@
 #include "CCNPPI.h"
 #include "JPEGEncoderNVJPEG.h"
 #include "test_utils.h"
+#include "nv_test_utils.h"
 #include "PipeLine.h"
 
 BOOST_AUTO_TEST_SUITE(ccnppi_tests)
+
 
 BOOST_AUTO_TEST_CASE(yuv411_I_1920x1080)
 {	
@@ -62,11 +64,13 @@ BOOST_AUTO_TEST_CASE(yuv411_I_1920x1080)
 	BOOST_TEST(outFrame->getMetadata()->getFrameType() == FrameMetadata::RAW_IMAGE_PLANAR);
 
 	Test_Utils::saveOrCompare("./data/testOutput/ccnppi_tests_yuv411_I_1920x1080_to_yuv444_1920x1080.raw", (const uint8_t *)outFrame->data(), outFrame->size(), 0);
-	p.term();
 	p.stop();
+	p.term();
+
 }
 
-BOOST_AUTO_TEST_CASE(mono_1920x1080_to_bgra_1920x1080)
+BOOST_AUTO_TEST_CASE(mono_1920x1080_to_bgra_1920x1080,
+* utf::precondition(if_compute_cap_supported()))
 {
 	auto width = 1920;
 	auto height = 1080;
@@ -112,8 +116,9 @@ BOOST_AUTO_TEST_CASE(mono_1920x1080_to_bgra_1920x1080)
 
 	Test_Utils::saveOrCompare("./data/testOutput/ccnppi_tests_mono_1920x1080_to_bgra_1920x1080.raw", (const uint8_t *)outFrame->data(), outFrame->size(), 0);
 
-	p.term();
 	p.stop();
+	p.term();
+
 }
 
 BOOST_AUTO_TEST_CASE(overlay_1920x960_BGRA_to_yuv420_1920x960)
@@ -161,8 +166,9 @@ BOOST_AUTO_TEST_CASE(overlay_1920x960_BGRA_to_yuv420_1920x960)
 	BOOST_TEST(outFrame->getMetadata()->getFrameType() == FrameMetadata::RAW_IMAGE_PLANAR);
 
 	Test_Utils::saveOrCompare("./data/testOutput/ccnppi_tests_overlay_1920x960_BGRA_to_yuv420_1920x960.raw", (const uint8_t *)outFrame->data(), outFrame->size(), 0);
-	p.term();
 	p.stop();
+	p.term();
+
 }
 
 BOOST_AUTO_TEST_CASE(yuv420_640x360_to_bgra_640x360)
@@ -210,11 +216,13 @@ BOOST_AUTO_TEST_CASE(yuv420_640x360_to_bgra_640x360)
 	BOOST_TEST(outFrame->getMetadata()->getFrameType() == FrameMetadata::RAW_IMAGE);
 
 	Test_Utils::saveOrCompare("./data/testOutput/ccnppi_tests_yuv420_640x360_to_bgra_640x360.raw", (const uint8_t *)outFrame->data(), outFrame->size(), 0);
-	p.term();
 	p.stop();
+	p.term();
+
 }
 
-BOOST_AUTO_TEST_CASE(yuv411_I_1920x1080__resize_to_jpg)
+BOOST_AUTO_TEST_CASE(yuv411_I_1920x1080__resize_to_jpg,
+* utf::precondition(if_compute_cap_supported()))
 {
 	auto width = 1920;
 	auto height = 1080;
@@ -265,8 +273,9 @@ BOOST_AUTO_TEST_CASE(yuv411_I_1920x1080__resize_to_jpg)
 	BOOST_TEST(outFrame->getMetadata()->getFrameType() == FrameMetadata::ENCODED_IMAGE);
 
 	Test_Utils::saveOrCompare("./data/testOutput/ccnppi_tests_yuv411_I_1920x1080_to_yuv444_960x540.jpg", (const uint8_t *)outFrame->data(), outFrame->size(), 0);
+	p.stop();
 	p.term();
-	p.stop();	
+
 }
 
 BOOST_AUTO_TEST_CASE(mono_1920x1080_to_yuv420_1920x1080)
@@ -315,8 +324,9 @@ BOOST_AUTO_TEST_CASE(mono_1920x1080_to_yuv420_1920x1080)
 	BOOST_TEST(outFrame->getMetadata()->getFrameType() == FrameMetadata::RAW_IMAGE_PLANAR);
 
 	Test_Utils::saveOrCompare("./data/testOutput/ccnppi_tests_mono_1920x1080_to_yuv420_1920x1080.raw", (const uint8_t *)outFrame->data(), outFrame->size(), 0);
-	p.term();
 	p.stop();
+	p.term();
+
 }
 
 BOOST_AUTO_TEST_CASE(mono_1920x960_to_yuv420_1920x960)
@@ -364,8 +374,9 @@ BOOST_AUTO_TEST_CASE(mono_1920x960_to_yuv420_1920x960)
 	BOOST_TEST(outFrame->getMetadata()->getFrameType() == FrameMetadata::RAW_IMAGE_PLANAR);
 
 	Test_Utils::saveOrCompare("./data/testOutput/ccnppi_tests_mono_1920x960_to_yuv420_1920x960.raw", (const uint8_t *)outFrame->data(), outFrame->size(), 0);
-	p.term();
 	p.stop();
+	p.term();
+
 }
 
 BOOST_AUTO_TEST_CASE(yuv420_640x360_to_yuv420_640x360)
@@ -414,8 +425,9 @@ BOOST_AUTO_TEST_CASE(yuv420_640x360_to_yuv420_640x360)
 	BOOST_TEST(outFrame->getMetadata()->getFrameType() == FrameMetadata::RAW_IMAGE_PLANAR);
 
 	Test_Utils::saveOrCompare("./data/testOutput/ccnppi_tests_yuv420_640x360_to_yuv420_640x360.raw", (const uint8_t *)outFrame->data(), outFrame->size(), 0);
+	p.stop();
 	p.term();
-	p.stop();	
+
 }
 
 BOOST_AUTO_TEST_CASE(perf, *boost::unit_test::disabled())
@@ -464,8 +476,9 @@ BOOST_AUTO_TEST_CASE(perf, *boost::unit_test::disabled())
 		copy2->step();
 		m3->pop();
 	}
+	p.stop();
 	p.term();
-	p.stop();	
+
 }
 
 BOOST_AUTO_TEST_SUITE_END()
